Revert "Feature/onboarding"
