+++ src/app/onboarding/SchoolButtonItem.tsx
@@ -1,12 +1,10 @@
 import styled from 'styled-components';
 import Text from '@/components/atoms/Text';
 import Link from 'next/link';
-import Image from 'next/image';
 import { SchoolButtonProps } from './interface';
 
 const Container = styled.div`
   position: relative;
-  display: block;
   height: 96px;
   margin: 0.75rem auto;
   background: linear-gradient(
@@ -14,17 +12,13 @@ const Container = styled.div`
     rgba(0, 0, 0, 0) 0%,
     rgba(0, 0, 0, 1) 100%
   );
-  border-radius: 0.5rem;
-  overflow: hidden;
-  text-decoration: none;
 `;
-
-const Background = styled(Image)`
+const Background = styled.img`
   position: absolute;
   width: 100%;
   height: 96px;
   object-fit: cover;
-  z-index: -1;
+  border-radius: 0.5rem;
 `;
 
 const GradientOverlay = styled.div`
@@ -42,7 +36,6 @@ const Title = styled(Text)`
   position: absolute;
   top: 0.75rem;
   left: 1.5rem;
-  color: white;
 `;
 
 export default function SchoolButtonItem({
@@ -51,9 +44,9 @@ export default function SchoolButtonItem({
   link,
 }: SchoolButtonProps) {
   return (
-    <Link href={link}>
-      <Container aria-label={`${title}로 이동`}>
-        <Background src={src} alt={title} fill />
+    <Link href={link} passHref>
+      <Container>
+        <Background src={src} alt={title} />
         <GradientOverlay />
         <Title variant="header3">{title}</Title>
       </Container>
