+++ src/app/photographer/[id]/screen.tsx
@@ -1,5 +1,3 @@
-/* eslint-disable import/no-extraneous-dependencies */
-
 'use client';
 
 import { useEffect, useState } from 'react';
